+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraInitRaceConditionsTests.swift
@@ -11,7 +11,7 @@ import XCTest
   import camera_avfoundation_objc
 #endif
 
-final class CameraCaptureSessionQueueRaceConditionTests: XCTestCase {
+final class CameraInitRaceConditionsTests: XCTestCase {
   private func createCameraPlugin() -> (CameraPlugin, DispatchQueue) {
     let captureSessionQueue = DispatchQueue(label: "io.flutter.camera.captureSessionQueue")
 
@@ -62,4 +62,36 @@ final class CameraCaptureSessionQueueRaceConditionTests: XCTestCase {
     XCTAssertNotNil(
       captureSessionQueue, "captureSessionQueue must not be nil after create method.")
   }
+
+  func testFlutterChannelInitializedWhenStartingImageStream() {
+    let (cameraPlugin, _captureSessionQueue) = createCameraPlugin()
+    let createExpectation = expectation(description: "create's result block must be called")
+
+    cameraPlugin.createCameraOnSessionQueue(
+      withName: "acamera",
+      settings: FCPPlatformMediaSettings.make(
+        with: .medium,
+        framesPerSecond: nil,
+        videoBitrate: nil,
+        audioBitrate: nil,
+        enableAudio: true
+      )
+    ) { result, error in
+      createExpectation.fulfill()
+    }
+
+    waitForExpectations(timeout: 30, handler: nil)
+
+    // Start stream and wait for its completion.
+    let startStreamExpectation = expectation(
+      description: "startImageStream's result block must be called")
+    cameraPlugin.startImageStream(completion: {
+      _ in
+      startStreamExpectation.fulfill()
+    })
+
+    waitForExpectations(timeout: 30, handler: nil)
+    XCTAssertEqual(cameraPlugin.camera?.isStreamingImages, true)
+  }
+
 }